fixup cyclic imports
