+++ qiskit/pulse/commands/pulse_decorators.py
@@ -18,31 +18,4 @@
 Pulse decorators.
 """
 
-import functools
-from typing import Callable
-
-import numpy as np
-
-from qiskit.pulse.exceptions import PulseError
-
-from .sample_pulse import SamplePulse
-
-
-def functional_pulse(func: Callable):
-    """A decorator for generating SamplePulse from python callable.
-
-    Args:
-        func: A function describing pulse envelope.
-    Raises:
-        PulseError: when invalid function is specified.
-    """
-    @functools.wraps(func)
-    def to_pulse(duration, *args, name=None, **kwargs):
-        """Return SamplePulse."""
-        if isinstance(duration, (int, np.integer)) and duration > 0:
-            samples = func(duration, *args, **kwargs)
-            samples = np.asarray(samples, dtype=np.complex128)
-            return SamplePulse(samples=samples, name=name)
-        raise PulseError('The first argument must be an integer value representing duration.')
-
-    return to_pulse
+from qiskit.pulse.pulse_lib.samplers.decorators import functional_pulse
